Consistency in identifier naming and typing.
We consistently use different property names for the underlying raw identifier values as Hibernate uses those as column names directly which causes clashes if multiple identifiers are used within a class.

We now consistently use String as the raw identifier type to avoid issues with UUID handling on H2 2.0 / Hibernate 5.